refactor(web): extract EditStudentForm from EditStudent's dialog
The metadata form (fields + validation + conflict-retry save) now lives in a
dialogless EditStudentForm so the roster detail modal can embed it without
nesting a second <dialog showModal>. EditStudent keeps its dialog shell around
the same form.
